+++ packages/sedk-mysql/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.1.0]  2023-12-07
+### Changed
+- Pump version to minor 1 since it is usable product
+- AliasedTable use Generic that extends Table instead of Table, to avoid build time error when using Column from table
+
 ## [0.0.4]  2023-11-16
 ### Added
 - support join, leftJoin, rightJoin and innerJoin steps.